% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_recode_question.R
\name{recode_primary_goal}
\alias{recode_primary_goal}
\title{recode_primary_goal}
\usage{
recode_primary_goal(goal)
}
\arguments{
\item{goal}{Numeric goal value.}
}
\value{
Primary goal label
}
\description{
Recode a primary goal code to a label.
}
